Separate and modify interpolation methods. (#20991)
