+++ framework/include/limiters/MinModLimiter.h
@@ -10,7 +10,7 @@
 #pragma once
 
 #include "Limiter.h"
-#include "MathFVUtils.h"
+#include "FVInterpolationUtils.h"
 
 namespace Moose
 {
