+++ framework/src/fvbcs/FVMatAdvectionFunctionBC.C
@@ -9,6 +9,7 @@
 
 #include "FVMatAdvectionFunctionBC.h"
 #include "Function.h"
+#include "FVInterpolationUtils.h"
 
 registerADMooseObject("MooseApp", FVMatAdvectionFunctionBC);
 
@@ -83,7 +84,6 @@ ADReal
 FVMatAdvectionFunctionBC::computeQpResidual()
 {
   ADReal flux_var_face;
-  ADRealVectorValue v_face;
 
   mooseAssert(
       _flux_variable_exact_solution,
@@ -91,23 +91,14 @@ FVMatAdvectionFunctionBC::computeQpResidual()
       "you suppress the flux_variable_exact_solution parameter in your derived class and forget to "
       "override computeQpResidual?");
 
-  Real flux_var_ghost = _flux_variable_exact_solution->value(
-      _t, 2. * _face_info->faceCentroid() - _face_info->elemCentroid());
+  Real flux_var_ghost = _flux_variable_exact_solution->value(_t, _face_info->faceCentroid());
   RealVectorValue v_ghost(
-      _vel_x_exact_solution.value(_t, 2. * _face_info->faceCentroid() - _face_info->elemCentroid()),
-      _vel_y_exact_solution ? _vel_y_exact_solution->value(
-                                  _t, 2. * _face_info->faceCentroid() - _face_info->elemCentroid())
-                            : 0,
-      _vel_z_exact_solution ? _vel_z_exact_solution->value(
-                                  _t, 2. * _face_info->faceCentroid() - _face_info->elemCentroid())
-                            : 0);
+      _vel_x_exact_solution.value(_t, _face_info->faceCentroid()),
+      _vel_y_exact_solution ? _vel_y_exact_solution->value(_t, _face_info->faceCentroid()) : 0,
+      _vel_z_exact_solution ? _vel_z_exact_solution->value(_t, _face_info->faceCentroid()) : 0);
 
-  interpolate(Moose::FV::InterpMethod::Average,
-              v_face,
-              _vel(makeElemArg(&_face_info->elem())),
-              v_ghost,
-              *_face_info,
-              true);
+  ADRealVectorValue v_face = Moose::FV::linearInterpolation(
+      _vel(makeElemArg(&_face_info->elem())), v_ghost, *_face_info, true);
 
   interpolate(_advected_interp_method,
               flux_var_face,
