+++ test/src/fvbcs/FVMatAdvectionOutflowBC.C
@@ -29,11 +29,6 @@ FVMatAdvectionOutflowBC::validParams()
                              advected_interp_method,
                              "The interpolation to use for the advected quantity. Options are "
                              "'upwind' and 'average', with the default being 'upwind'.");
-  params.addParam<bool>(
-      "use_ssf",
-      false,
-      "Whether to evaluate functors using a single-sided face argument. If this is false, then we "
-      "will evaluate the functors using ElemFromFace arguments.");
   return params;
 }
 
@@ -41,8 +36,7 @@ FVMatAdvectionOutflowBC::FVMatAdvectionOutflowBC(const InputParameters & params)
   : FVFluxBC(params),
     _vel(getFunctor<ADRealVectorValue>("vel")),
     _adv_quant(getFunctor<ADReal>(isParamValid("advected_quantity") ? "advected_quantity"
-                                                                    : variable().name())),
-    _use_ssf(getParam<bool>("use_ssf"))
+                                                                    : variable().name()))
 {
   using namespace Moose::FV;
 
@@ -59,33 +53,9 @@ FVMatAdvectionOutflowBC::FVMatAdvectionOutflowBC(const InputParameters & params)
 ADReal
 FVMatAdvectionOutflowBC::computeQpResidual()
 {
-  using namespace Moose::FV;
-
-  ADRealVectorValue v;
-  ADReal adv_quant_boundary;
-
-  if (_use_ssf)
-  {
-    const auto ssf = singleSidedFaceArg();
-    v = _vel(ssf);
-    adv_quant_boundary = _adv_quant(ssf);
-  }
-  else
-  {
-    const auto elem_face = elemFromFace();
-    const auto neighbor_face = neighborFromFace();
-
-    // Currently only Average is supported for the velocity
-    interpolate(InterpMethod::Average, v, _vel(elem_face), _vel(neighbor_face), *_face_info, true);
-
-    interpolate(_advected_interp_method,
-                adv_quant_boundary,
-                _adv_quant(elem_face),
-                _adv_quant(neighbor_face),
-                v,
-                *_face_info,
-                true);
-  }
+  const auto ssf = singleSidedFaceArg();
+  ADRealVectorValue v = _vel(ssf);
+  ADReal adv_quant_boundary = _adv_quant(ssf);
 
   return _normal * v * adv_quant_boundary;
 }
